+++ failgood/src/failgood/internal/execution/ContextStateCollector.kt
@@ -1,4 +1,4 @@
-package failgood.internal.execution.context
+package failgood.internal.execution
 
 import failgood.*
 import failgood.dsl.TestFunction
