+++ failgood/src/failgood/internal/execution/StaticContextExecutionConfig.kt
@@ -1,4 +1,4 @@
-package failgood.internal.execution.context
+package failgood.internal.execution
 
 import failgood.ExecutionListener
 import failgood.NullExecutionListener
